Rename Logistic Regression for 1 parameter to Logistic Regression for 1 para as Sigmoid
Target of this program is to predict the defective of hardware by measuring parameter of sensor. For example, temperature vs failed part due to heat rejection. 

Generally, Products are tuned to avoid such a failure during the developing term.However , it may be not optimum accidently. 

This program can correct parameter by field actual data.